oneplus2: Huge updates to script and support uploading to AFH
Credits- MSF-Jarvis

+++ oneplus2-build.sh
@@ -1,13 +1,28 @@
-BUILD_START=$(date +"%s")
+#!/bin/bash
+
+# Declaring colours
 blue='\033[0;34m'
 cyan='\033[0;36m'
 yellow='\033[0;33m'
 red='\033[0;31m'
 nocol='\033[0m'
+
+# Variables for build
 TOOLCHAIN_DIR=~/toolchain/bin
 KERNEL_DIR=~/kernel
 KERN_IMG=$KERNEL_DIR/arch/arm64/boot/Image.gz-dtb
 
+# Var declaration
+bool=N
+
+# DarkBeast kernel details 
+KERNEL_NAME="DarkBeast-Kernel"
+VER="v1"
+DEVICE="oneplus2"
+DATE="$(date +"%Y%m%d")"
+FINAL_ZIP="$KERNEL_NAME-$VER-$DEVICE-$DATE"
+
+# Usefull components for build
 export ARCH=arm64
 export KBUILD_BUILD_USER="The_DarkBeast"
 export KBUILD_BUILD_HOST="Weed-Machine"
@@ -15,27 +30,61 @@ STRIP="~/toolchain/bin/aarch64-"
 export CCOMPILE=$CROSS_COMPILE
 export CROSS_COMPILE=aarch64-
 export PATH=$PATH:~/toolchain/bin
+DATE_START=$(date +"%s")
 
 cd ~/kernel
 echo -e "Making Config"
 make oneplus2_defconfig
 echo -e "Starting Build"
-echo -e "$blue**************************************************************************** $nocal"
+echo -e "$blue**************************************************************************** $nocol"
 echo "                    "
-echo "                                   $yellowCompiling DarkBeast Kernel $nocal             "
+echo "                                   Compiling DarkBeast Kernel             "
 echo "                    "
-echo -e "$blue**************************************************************************** $nocal"
+echo -e "$blue**************************************************************************** $nocol"
 make -j32
+
 if ! [ -a $KERN_IMG ];
 then
-echo -e "$red Kernel Compilation failed! Fix the errors! $nocol"
-exit 1
+    echo -e "$red Kernel Compilation failed! Fix the errors! $nocol"
+    exit 1
 fi
-cd ..
-echo -e "Moving shiz"
-rm -rf ~/zipper/tools/Image.gz-dtb
-rm -rf ~/zipper/DarkBeast*
-cp ~/kernel/arch/arm64/boot/Image.gz-dtb ~/zipper/Image.gz-dtb
-cd ~/zipper
-zip -r DarkBeast-Kernel-v2-oneplus2-$(date +"%Y%m%d").zip *
-echo -e "Done"
+   DATE_END=$(date +"%s")
+   DIFF=$(($DATE_END - $DATE_START))
+   echo "It took $(($DIFF / 60)) minute(s) and $(($DIFF % 60)) seconds to build the kernel"
+   echo " "
+   cd ..
+
+echo -e " "
+echo -e "Do you want to build the zip y/N"
+read bool
+
+if [ $bool = N ]
+then
+    echo -e "$(cyan) Kernel has been compiled succesfully $(nocol)"
+    exit 1
+
+else
+    echo -e "Moving required components to zipper"
+    rm -rf ~/zipper/tools/Image.gz-dtb
+    rm -rf ~/zipper/DarkBeast*
+    cp ~/kernel/arch/arm64/boot/Image.gz-dtb ~/zipper/Image.gz-dtb
+    cd ~/zipper
+    echo -e " "
+    echo -e "$(yellow) Zipping all Contents $(nocol)"
+    zip -r $FINAL_ZIP.zip *
+    echo -e " "
+    echo -e "Zipped all the contents"
+    echo -e "$(cyan) $FINAL_ZIP $(nocol)"
+fi
+
+bool=N
+echo -e "Do you want to upload the zip y/N"
+read bool
+
+if [ $bool = y ]
+then
+    wput ftp://${AFH_CREDENTIALS}@uploads.androidfilehost.com/ ${FINAL_VER}.zip
+    echo -e "$FINAL_ZIP has been uploaded to your AndroidFileHost account"
+    echo -e "Build Done"
+fi
+    echo -e "Build Done"
